refactor(Handler): move class in folders
